requirements.txt
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
M2Crypto==0.37.1
numpy==1.20.3
opencv-python==4.5.2.52
parameterized==0.8.1
Pillow==8.2.0
pycparser==2.20
pycryptodome==3.10.1
pyOpenSSL==20.0.1
pyscard==2.0.0
PySimpleGUI==4.43.0
six==1.16.0
